requirements.txt
# absl-py==1.3.0
# altair==4.2.0
# anyio==3.6.2
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.2.3
# asttokens==2.2.1
# astunparse==1.6.3
# attrs==22.1.0
# backcall==0.2.0
# beautifulsoup4==4.11.1
# bleach==5.0.1
# blinker==1.5
# cachetools==5.2.0
# certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==2.1.1
# click==8.1.3
# colorama==0.4.6
# comm==0.1.2
# commonmark==0.9.1
# debugpy==1.6.4
# decorator==5.1.1
# defusedxml==0.7.1
# entrypoints==0.4
# executing==1.2.0
# fastjsonschema==2.16.2
# flatbuffers==22.12.6
# fqdn==1.5.1
# gast==0.4.0
# gitdb==4.0.10
# GitPython==3.1.29
# google-auth==2.15.0
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.51.1
# h5py==3.7.0
# idna==3.4
# importlib-metadata==5.2.0
# ipykernel==6.19.4
# ipython==8.7.0
# ipython-genutils==0.2.0
# ipywidgets==8.0.3
# isoduration==20.11.0
# jedi==0.18.2
# Jinja2==3.1.2
# joblib==1.2.0
# jsonpointer==2.3
# jsonschema==4.17.3
# jupyter==1.0.0
# jupyter-console==6.4.4
# jupyter-events==0.5.0
# jupyter_client==7.4.8
# jupyter_core==5.1.0
# jupyter_server==2.0.2
# jupyter_server_terminals==0.4.3
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==3.0.4
# keras==2.10.0
# Keras-Preprocessing==1.1.2
# libclang==14.0.6
# Markdown==3.4.1
# MarkupSafe==2.1.1
# matplotlib-inline==0.1.6
# mistune==2.0.4
# nbclassic==0.4.8
# nbclient==0.7.2
# nbconvert==7.2.7
# nbformat==5.7.1
# nest-asyncio==1.5.6
# notebook==6.5.2
# notebook_shim==0.2.2
# numpy==1.24.0
# oauthlib==3.2.2
# opt-einsum==3.3.0
# packaging==22.0
# pandas==1.5.2
# pandocfilters==1.5.0
# parso==0.8.3
# pickleshare==0.7.5
# Pillow==9.3.0
# platformdirs==2.6.0
# prometheus-client==0.15.0
# prompt-toolkit==3.0.36
# protobuf==3.19.6
# psutil==5.9.4
# pure-eval==0.2.2
# pyarrow==10.0.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.21
# pydeck==0.8.0
# Pygments==2.13.0
# Pympler==1.0.1
# pyrsistent==0.19.2
# python-dateutil==2.8.2
# python-json-logger==2.0.4
# pytz==2022.7
# pytz-deprecation-shim==0.1.0.post0
# pywin32==305
# pywinpty==2.0.9
# PyYAML==6.0
# pyzmq==24.0.1
# qtconsole==5.4.0
# QtPy==2.3.0
# requests==2.28.1
# requests-oauthlib==1.3.1
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rich==12.6.0
# rsa==4.9
# scikit-learn==1.2.0
# scipy==1.9.3
# semver==2.13.0
# Send2Trash==1.8.0
# six==1.16.0
# smmap==5.0.0
# sniffio==1.3.0
# soupsieve==2.3.2.post1
# split-folders==0.5.1
# stack-data==0.6.2
# streamlit==1.16.0
# streamlit-lottie==0.0.3
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.1
# tensorflow-estimator==2.10.0
# tensorflow-io-gcs-filesystem==0.29.0
# termcolor==2.1.1
# terminado==0.17.1
# threadpoolctl==3.1.0
# tinycss2==1.2.1
# toml==0.10.2
# toolz==0.12.0
# tornado==6.2
# traitlets==5.8.0
# typing_extensions==4.4.0
# tzdata==2022.7
# tzlocal==4.2
# uri-template==1.2.0
# urllib3==1.26.13
# validators==0.20.0
# watchdog==2.2.0
# wcwidth==0.2.5
# webcolors==1.12
# webencodings==0.5.1
# websocket-client==1.4.2
# Werkzeug==2.2.2
# widgetsnbextension==4.0.4
# wrapt==1.14.1
# zipp==3.11.0
absl-py==1.3.0
altair==4.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
flatbuffers==22.12.6
fqdn==1.5.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.2.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.2
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
split-folders==0.5.1
SQLAlchemy==1.4.45
stack-data==0.6.2
streamlit==1.16.0
streamlit-lottie==0.0.3
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.4
wrapt==1.14.1
zipp==3.11.0
